move mp4 recording to util
